+++ python/dgl/distributed/__init__.py
@@ -9,4 +9,4 @@ from .rpc_server import start_server
 from .rpc_client import connect_to_server, finalize_client, shutdown_servers
 from .kvstore import KVServer, KVClient
 from .server_state import ServerState
-from .sampling import sample_neighbors, in_subgraph
+from .graph_services import sample_neighbors, in_subgraph
